requirements.txt
Babel==2.3.4
Django==1.9.7
django-bootstrap3==7.0.1
django-phonenumber-field==1.1.0
phonenumberslite==7.4.2
pytz==2016.4
